some cleanup and split up tests
