--- backend/quickstart/.gitignore
@@ -1,1 +1,0 @@
-@cds-models
