--- backend/quickstart/.vscode/tasks.json
@@ -1,10 +1,0 @@
-{
-  "version": "2.0.0",
-  "tasks": [
-    {
-      "type": "shell",
-      "label": "cds watch",
-      "command": "cds watch",
-    }
-  ]
-}
