+++ backend/quickstart/test/document-service/assets/reference-architecture-generative-ai-basic.drawio
