+++ backend/quickstart/test/document-service/main.test.js
@@ -139,7 +139,7 @@ async function createDocumentViaAction(
   return json;
 }
 
-describe('QuickstartService API (integration)', function () {
+describe('Document Service (integration)', function () {
   // Allow some time for local DB operations.
   this.timeout(20_000);
 
@@ -300,164 +300,5 @@ describe('QuickstartService API (integration)', function () {
     assert.equal(downloadRes.status, 200);
     assert.match(downloadRes.headers.get('content-type') || '', /image\/png/i);
     assert.equal(downloadedBytes.compare(sourceBytes), 0, 'Downloaded PNG bytes must match uploaded content');
-  });
-});
-
-describe('QuickstartService API (authorization)', function () {
-  this.timeout(20_000);
-
-  let alice;
-  let bob;
-  let fred;
-
-  before(async () => {
-    alice = await createCsrfSession('alice');
-    bob = await createCsrfSession('bob');
-    fred = await createCsrfSession('fred');
-  });
-
-  it('does not expose Users directly', async () => {
-    const { res } = await requestJson(alice, '/Users');
-    assert.equal(res.status, 403);
-  });
-
-  it('returns author and contributor usernames via Documents expand', async () => {
-    const created = await createDocumentViaAction(alice, {
-      title: `Expand Users ${crypto.randomUUID().slice(0, 8)}`,
-      contributorsUsernames: ['bob'],
-    });
-
-    const { res, json } = await requestJson(alice, `/Documents(${created.ID})?$expand=author,contributors($expand=user)`);
-    assert.equal(res.status, 200);
-
-    assert.equal(json?.author?.username, 'alice', 'Expected author username in expanded document');
-
-    const contributorUsernames = (json?.contributors || [])
-      .map((entry) => entry?.user?.username)
-      .filter(Boolean);
-    assert.ok(contributorUsernames.includes('bob'), 'Expected contributor username in expanded document');
-  });
-
-  it('rejects direct POST to Documents', async () => {
-    const { res } = await requestJson(alice, '/Documents', {
-      method: 'POST',
-      body: {
-        title: 'Should be rejected',
-        description: 'Must use createNewDocument action',
-        author_ID: 'A0B1C2D3-1111-4A11-8111-111111111111',
-        editorState: '{}',
-      },
-    });
-
-    assert.ok([403, 405].includes(res.status), `Expected direct POST /Documents to be rejected, got ${res.status}`);
-  });
-
-  it('enforces Tags as read-only', async () => {
-    const tagCode = `AUTH-TAG-${crypto.randomUUID().slice(0, 8)}`.toUpperCase();
-
-    const { res } = await requestJson(alice, '/Tags', {
-      method: 'POST',
-      body: {
-        code: tagCode,
-        label: 'Authorization Test Tag',
-        description: 'Should not be creatable via API',
-      },
-    });
-
-    assert.ok(res.status >= 400, `Expected write to @readonly Tags to fail, got ${res.status}`);
-  });
-
-  it('allows contributor read but blocks non-author writes', async () => {
-    const { res: createRes, json: created } = await requestJson(alice, '/createNewDocument', {
-      method: 'POST',
-      body: {
-        title: `Auth Test Doc ${crypto.randomUUID().slice(0, 8)}`,
-        description: 'authorization test',
-        parentId: null,
-        tags: ['agents'],
-        contributorsUsernames: ['bob'],
-        editorState: '{}',
-      },
-    });
-
-    assert.equal(createRes.status, 200);
-    const documentId = created?.ID;
-    assert.ok(documentId, 'Expected created document ID');
-
-    const { res: contributorReadRes } = await requestJson(bob, `/Documents(${documentId})`);
-    assert.equal(contributorReadRes.status, 200, 'Contributor should be able to read document');
-
-    const { res: unrelatedReadRes } = await requestJson(fred, `/Documents(${documentId})`);
-    assert.ok(
-      [403, 404].includes(unrelatedReadRes.status),
-      `Unrelated user should not read document, got ${unrelatedReadRes.status}`
-    );
-
-    const { res: contributorPatchRes } = await requestJson(bob, `/Documents(${documentId})`, {
-      method: 'PATCH',
-      body: {
-        description: 'bob attempted update',
-      },
-    });
-    assert.ok(
-      [403, 404].includes(contributorPatchRes.status),
-      `Contributor should not update document, got ${contributorPatchRes.status}`
-    );
-
-    const assetId = uuid();
-    const { res: contributorAssetWriteRes } = await requestJson(bob, '/DocumentAssets', {
-      method: 'POST',
-      body: {
-        ID: assetId,
-        mediaType: 'image/png',
-        filename: 'auth-test.png',
-        document_ID: documentId,
-      },
-    });
-    assert.ok(
-      [403, 404].includes(contributorAssetWriteRes.status),
-      `Contributor should not create document assets, got ${contributorAssetWriteRes.status}`
-    );
-
-    const { res: contributorTagWriteRes } = await requestJson(bob, '/DocumentTags', {
-      method: 'POST',
-      body: {
-        document_ID: documentId,
-        tag_code: 'data',
-      },
-    });
-    assert.ok(
-      [400, 403, 404].includes(contributorTagWriteRes.status),
-      `Contributor should not create document tags, got ${contributorTagWriteRes.status}`
-    );
-
-    const { res: contributorContribWriteRes } = await requestJson(bob, '/setDocumentContributors', {
-      method: 'POST',
-      body: {
-        documentId,
-        contributorsUsernames: ['bob', 'fred'],
-      },
-    });
-    assert.equal(contributorContribWriteRes.status, 403, 'Contributor should not change contributor list');
-  });
-
-  it('returns only documents visible to the current user in collection reads', async () => {
-    const visibleToBob = await createDocumentViaAction(alice, {
-      title: `Visible-To-Bob ${crypto.randomUUID().slice(0, 8)}`,
-      contributorsUsernames: ['bob'],
-    });
-
-    const hiddenFromBob = await createDocumentViaAction(alice, {
-      title: `Hidden-From-Bob ${crypto.randomUUID().slice(0, 8)}`,
-      contributorsUsernames: [],
-    });
-
-    const { res, json } = await requestJson(bob, '/Documents');
-    assert.equal(res.status, 200);
-    assert.ok(json && Array.isArray(json.value), 'Expected OData collection response');
-
-    const returnedIds = new Set(json.value.map((d) => d.ID));
-    assert.ok(returnedIds.has(visibleToBob.ID), 'Expected bob-visible document to be in collection result');
-    assert.ok(!returnedIds.has(hiddenFromBob.ID), 'Expected bob-hidden document to be filtered out from collection result');
   });
 });
